feat: publish Archive Test fixture pairs to folders without overwriting
Implements #838 (slice 04/12 of #834): ArchiveTestRequest validates
selection, duplicates, and a fresh destination resolved under the
working directory; ArchiveTestSuiteGenerator stages flat fixtureId.zip +
fixtureId.json pairs in an owned GUID sibling directory with exclusive
creates, validates semantics before publication, guards the final
non-overwriting rename (POSIX rename silently replaces empty
destinations), and on failure removes only the owned staging directory.
14 filesystem tests; no CLI yet.

+++ src/Zipper.Tests/ArchiveTests/ArchiveTestCaseSchemaTests.cs
@@ -182,7 +182,7 @@ public class ArchiveTestCaseSchemaTests
 
     private static bool IsBoundedHex(string value, int maxBytes) =>
         value.Length % 2 == 0
-        && value.Length <= maxBytes * 2
+        && value.Length <= (long)maxBytes * 2
         && IsLowercaseHex(value);
 
     private static bool IsLowercaseHex(string value) =>
